+++ src/Patches/ZNetPatches.cs
@@ -30,13 +30,59 @@ namespace DiscordConnector.Patches
         {
             private static void Postfix(ref bool server, ref bool openServer, ref bool publicServer, ref string serverName, ref string password, ref World world)
             {
-                DiscordApi.SendMessage("Loaded server info");
-                Plugin.StaticServerSetup.Add(Plugin.ServerSetup.IsServer, server);
-                Plugin.StaticServerSetup.Add(Plugin.ServerSetup.IsOpenServer, openServer);
-                Plugin.StaticServerSetup.Add(Plugin.ServerSetup.IsPublicServer, publicServer);
-                Plugin.StaticServerInfo.Add(Plugin.ServerInfo.WorldName, world.m_name);
-                Plugin.StaticServerInfo.Add(Plugin.ServerInfo.WorldSeed, $"{world.m_seed}");
-                Plugin.StaticServerInfo.Add(Plugin.ServerInfo.WorldSeedName, world.m_seedName);
+                if (Plugin.StaticServerSetup.ContainsKey(Plugin.ServerSetup.IsServer))
+                {
+                    Plugin.StaticServerSetup[Plugin.ServerSetup.IsServer] = server == true;
+                }
+                else
+                {
+                    Plugin.StaticServerSetup.Add(Plugin.ServerSetup.IsServer, server == true);
+                }
+
+                if (Plugin.StaticServerSetup.ContainsKey(Plugin.ServerSetup.IsOpenServer))
+                {
+                    Plugin.StaticServerSetup[Plugin.ServerSetup.IsOpenServer] = openServer == true;
+                }
+                else
+                {
+                    Plugin.StaticServerSetup.Add(Plugin.ServerSetup.IsOpenServer, openServer == true);
+                }
+
+                if (Plugin.StaticServerSetup.ContainsKey(Plugin.ServerSetup.IsPublicServer))
+                {
+                    Plugin.StaticServerSetup[Plugin.ServerSetup.IsPublicServer] = publicServer == true;
+                }
+                else
+                {
+                    Plugin.StaticServerSetup.Add(Plugin.ServerSetup.IsPublicServer, publicServer == true);
+                }
+
+                if (Plugin.StaticServerInfo.ContainsKey(Plugin.ServerInfo.WorldName))
+                {
+                    Plugin.StaticServerInfo[Plugin.ServerInfo.WorldName] = $"{world.m_name}";
+                }
+                else
+                {
+                    Plugin.StaticServerInfo.Add(Plugin.ServerInfo.WorldName, $"{world.m_name}");
+                }
+
+                if (Plugin.StaticServerInfo.ContainsKey(Plugin.ServerInfo.WorldSeed))
+                {
+                    Plugin.StaticServerInfo[Plugin.ServerInfo.WorldSeed] = $"{world.m_seed}";
+                }
+                else
+                {
+                    Plugin.StaticServerInfo.Add(Plugin.ServerInfo.WorldSeed, $"{world.m_seed}");
+                }
+
+                if (Plugin.StaticServerInfo.ContainsKey(Plugin.ServerInfo.WorldSeedName))
+                {
+                    Plugin.StaticServerInfo[Plugin.ServerInfo.WorldSeedName] = $"{world.m_seedName}";
+                }
+                else
+                {
+                    Plugin.StaticServerInfo.Add(Plugin.ServerInfo.WorldSeedName, $"{world.m_seedName}");
+                }
             }
         }